File structure and some data containers
Started with getting some scripts/prefab/material folders ready and getting a little bit of structure in the project window.

I have made a few scripts that can hold necessary data. Level Data is the data that gets send to the server (I still have to find a way to keep track of the level stats)

VoxelData holds the grid position of one voxel and the final color (when the puzzle completes).

Voxel script is used during gameplay. It has a world space canvas on each side for the nonogram clues. Also holds the state of the voxel and the different colors/materials needed.

Level creation script is used to turn the data in leveldata into a playable puzzle.

I will be adding some more logic in the GridData class and start working on a level editor tomorrow.