Expanded the gridData script and begin towards level editor
Grid data now holds grid size, all the grids that are filled and that are empty. With that information you should be able to fully build a nonogram puzzle. We only need a way to fill in that data and that is what the level editor is for.

Made some enums and ways to customize the level editor. It only needs actual functionality, but that is the hard part.

In the voxelinteraction script you can now click on a voxel and enter it in 1 of 4 states: drag, paint, mark and destroy. The state can be changed with a button later