Added cinemachine for orbit camera control and worked a little bit on level editor
I wanted to use a box collider and have a script that shoots a raycast through the collider and get the second hitpoint so that you are in the boxcollider. But unity doesnt support hitting the same collider twice with 1 raycast.
I am currently working on making a script that makes a "Bounding box" out of 6 planes that scale with gridsize so that way the colliders are seperate and you can get "Inside" the box to place voxels