Bounding box scales with gridsize and voxels can be placed inside the bounding box
you can place some voxels outside of the bounding box, trying to find a solution for that. I also need to link the placed voxels with a data container.

Also using unlit shaders now for the voxels and bounding box.

Had some fps issues within the editor and play mode. Whenever using my mouse fps would drop, apparently is was because of high mouse polling rate so should be more carefull about that.