New scene for puzzles, and refactored level editor
you can now undo and redo your actions in the level editor, whenever you shrink the gridsize, it properly deletes all the voxels outside the grid. still have to fixx the undo for changing gridsize.

made a new scene and set a few things up but it is not done yet.

Made a scriptable object that holds the leveldata to persist between scenes