Server status debugging ingame and changed some of the main puzzle scripts
I kinda messed up and forgot to make a diagram for the flow of the project, I think I have a lot of duplicate/unnecessary logic that just makes managing this project a lot harder.

in this commit I added some sever debugging in game so people know when the server is online or not. Also made some changes to the nonogram clue generator. Instead of using keys in a dictionary using a 3d array.

Still need to add some sort of layer slicing otherwise some levels are impossible