Finishing touches
FIxxed the calculation errors in clue generator, when creating the lvel it would generate the voxel 0,0,0 in the corner and not in the center which caused some issues but with some offset fixxes it works.

Clue genereator is still kinda sloppy and sometimes does nto work the 100%

added some extra UI functionality for the full gameplayloop. Also fully implemented the color picker