Finishing touches
Got some feedback on the material, so I changed it back to lit and added directional light in my scenes.

fixxed auto font size for the voxel clues.

Fixxed offset calculation in the nonogram clue generator it properly show the clues now even on different gridsizes.

Added a small hint in the main menu screen to explain a little what a nonogram is